fix(cancel-ride): Update logic for canceling rides
- Improved error handling for ride cancellation
- Updated UI to reflect canceled ride status
