+++ src/pages/driver/RideRequest.tsx
@@ -1,4 +1,5 @@
 import LoadingSpinner from "@/components/layout/LoadingSpinner";
+import LocationDisplay from "@/components/modules/RideRequest/LocationDisplay";
 import RideStepper from "@/components/modules/RideRequest/RideStepper";
 import { Badge } from "@/components/ui/badge";
 import {
@@ -10,56 +11,7 @@ import {
 } from "@/components/ui/card";
 import { useGetAllRidesQuery } from "@/redux/features/ride.api";
 import type { IRide, RideStatus } from "@/types/ride-type";
-import { getLocationName } from "@/utils/location";
-import { Calendar, DollarSign, MapPin } from "lucide-react";
-import { useEffect, useState } from "react";
-
-// Component to display a location with loading state
-const LocationDisplay = ({
-  lat,
-  lng,
-  label,
-}: {
-  lat: number;
-  lng: number;
-  label: string;
-}) => {
-  const [locationName, setLocationName] = useState<string | null>(null);
-  const [isLoading, setIsLoading] = useState(true);
-
-  useEffect(() => {
-    const fetchLocation = async () => {
-      setIsLoading(true);
-      try {
-        const name = await getLocationName(lat, lng);
-        setLocationName(name);
-      } catch (error) {
-        console.error("Error fetching location:", error);
-        setLocationName(`${lat.toFixed(4)}, ${lng.toFixed(4)}`);
-      } finally {
-        setIsLoading(false);
-      }
-    };
-
-    fetchLocation();
-  }, [lat, lng]);
-
-  return (
-    <div className="flex items-center gap-3">
-      <MapPin className="text-muted-foreground h-5 w-5" />
-      <div>
-        <p className="text-muted-foreground text-sm">{label}</p>
-        {isLoading ? (
-          <p className="text-muted-foreground font-medium">
-            Loading address...
-          </p>
-        ) : (
-          <p className="font-medium">{locationName}</p>
-        )}
-      </div>
-    </div>
-  );
-};
+import { Calendar, DollarSign } from "lucide-react";
 
 const RideRequest = () => {
   const { data: rides, isLoading, isFetching } = useGetAllRidesQuery({});
